sandbox/poc/requirements.txt
# Core dependencies
streamlit>=1.28.0
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
